feat(aapd-861): move api client to common
