+++ packages/common/src/client/appeals-api-error.js
@@ -1,0 +1,20 @@
+/**
+ * @class Api Client for v2 urls in appeals-service-api
+ */
+class AppealsApiError extends Error {
+	/**
+	 * @param {string} message
+	 * @param {string|number} code
+	 * @param {Array.<string>} [errors]
+	 */
+	constructor(message, code, errors) {
+		super(message);
+		this.name = 'AppealsApiError';
+		/** @type {string|number} */
+		this.code = code;
+		/** @type {Array.<string>|undefined} */
+		this.errors = errors;
+	}
+}
+
+module.exports = AppealsApiError;
