+++ packages/common/src/config.js
@@ -9,5 +9,13 @@
 module.exports = {
 	logger: {
 		level: process.env.LOGGER_LEVEL || 'info'
+	},
+	apis: {
+		appealsApi: {
+			url: process.env.APPEALS_SERVICE_API_URL,
+			timeout: process.env.APPEALS_SERVICE_API_TIMEOUT
+				? parseInt(process.env.APPEALS_SERVICE_API_TIMEOUT, 10)
+				: 10000
+		}
 	}
 };
